[ts-5] Add a bag class.
A bag is an extension of a map. It allows multiple values to be set
for the same key. This a primary value, this is the value that will
be see and manipulated when using the bag as a map. It also has
secondary values, these are values that can be seen when using extra
methods provided by the bag.

Iterators and values will also show the extra secondary values.