feature: Add capturing errors when creating values
Some libraries or standard classes throw exceptions when constructed,
making ti hard to create them directly on a class:

.. code-block:: java

    class Example {
        private final URI uri = new URI ("...");
	public Example () throws Exception {}
    }

The constructor is required. We need to specify that it can throw
exceptions. Which makes it uglier that it needs to be to create the uri
(keep in mind that you can use ``URI.create`` in more recent java
versions, which would make the above a bad example).

Instead we could explicitly turn the exception into a runtime exception:

.. code-block:: java

    class Example {
        private final URI uri = capture (() -> new URI ("..."));
    }