improve: Support computing dependent option values
When using the compute method while computing an option, the
implementation tries to insert both results in the map. It does this in
reverse order. When the computation of the option completes the map has
been modified when the original option was computed. For a hash map this
causes an error.

This can be fix by either
- computing the value without using `computeIfAbsent ()`, and only
  afterwards inserting the result. This way only one modification is
  happening concurrently.

- Switching to a concurrent implementation which can handle these
  situations.

The second option has the advantage that the config class should also be
usable in concurrent situations, while leaving the implementation
readable.